Move system-specific logic to system catalog, out of base grid
