+++ src/app/modules/it-systems/it-system-catalog/it-system-catalog.component.html
@@ -28,4 +28,5 @@
   [entityType]="'it-system'"
   (stateChange)="stateChange($event)"
   (rowIdSelect)="rowIdSelect($event)"
+  (checkboxChange)="handleSystemUsageChange($event)"
 ></app-grid>
